Replacing fabs with fabsf

+++ src/graph/graph.c
@@ -2726,7 +2726,7 @@ static void update_work_queue_nodes_acc(size_t * __restrict__ num_work_items_nod
 #pragma omp parallel for default(none) shared(current_index, num_work_items_nodes, work_queue_scratch, convergence, work_queue_nodes, node_states, node_states_previous, node_states_current) private(i)
 #pragma acc kernels copyin(work_queue_nodes[0:num_vertices], node_states[0:num_vertices], node_states_current[0:num_vertices], node_states_previous[0:num_vertices]) copyout(work_queue_scratch[0:num_vertices])
 	for(i = 0; i < *num_work_items_nodes; ++i) {
-		if(fabs(node_states_current[work_queue_nodes[i]] - node_states_previous[work_queue_nodes[i]]) >= convergence) {
+		if(fabsf(node_states_current[work_queue_nodes[i]] - node_states_previous[work_queue_nodes[i]]) >= convergence) {
 #pragma omp critical
 #pragma acc atomic capture
 			{
@@ -3199,7 +3199,7 @@ static void update_work_queue_edges_acc(size_t * __restrict__ num_work_items_edg
 #pragma omp parallel for default(none) shared(current_index, num_work_items_edges, work_queue_scratch, convergence, work_queue_edges, current_state, previous_state) private(i)
 #pragma acc kernels copyin(work_queue_edges[0:num_edges], previous_state[0:num_edges], current_state[0:num_edges]) copyout(work_queue_scratch[0:num_edges])
 	for(i = 0; i < *num_work_items_edges; ++i) {
-		if(fabs(current_state[work_queue_edges[i]] - previous_state[work_queue_edges[i]]) >= convergence) {
+		if(fabsf(current_state[work_queue_edges[i]] - previous_state[work_queue_edges[i]]) >= convergence) {
 #pragma omp critical
 #pragma acc atomic capture
 			{
